// src/service-orders/enums/service-order-operative-status.enum.ts
export enum ServiceOrderOperativeStatus {
  ABIERTA = 'ABIERTA',
  EN_PROCESO = 'EN_PROCESO',
  CANCELACION_SOLICITADA = 'CANCELACION_SOLICITADA',
  LISTA_PARA_ENTREGA = 'LISTA_PARA_ENTREGA',
  ENTREGA_PARCIAL = 'ENTREGA_PARCIAL',
  ENTREGADA = 'ENTREGADA',
  CANCELADA = 'CANCELADA',
  CERRADA_SIN_SOLUCION = 'CERRADA_SIN_SOLUCION',
}