// src/service-orders/enums/service-order-technical-status.enum.ts
export enum ServiceOrderTechnicalStatus {
  PENDIENTE_ASIGNACION = 'PENDIENTE_ASIGNACION',
  ASIGNADA = 'ASIGNADA',
  EN_DIAGNOSTICO = 'EN_DIAGNOSTICO',
  DIAGNOSTICADA = 'DIAGNOSTICADA',
  PENDIENTE_DEFINICION_COMERCIAL = 'PENDIENTE_DEFINICION_COMERCIAL',
  AUTORIZADA_PARA_EJECUCION = 'AUTORIZADA_PARA_EJECUCION',
  EN_EJECUCION = 'EN_EJECUCION',
  BLOQUEADA = 'BLOQUEADA',
  ESPERANDO_REPUESTOS_O_TERCERO = 'ESPERANDO_REPUESTOS_O_TERCERO',
  RESUELTA = 'RESUELTA',
  SIN_SOLUCION = 'SIN_SOLUCION',
}